refactor(middleware): extract permission handling to base class
Makes it possible to use the same logic for Stemmarest and Stemweb
