+++ middleware/stemmaweb_middleware/stemmarest/permissions/declarations/users.py
@@ -14,7 +14,9 @@ from stemmaweb_middleware.stemmarest.permissions.service import (
 )
 
 
-def config(service: StemmarestPermissionService, args: PermissionArguments):
+def config(
+    service: StemmarestPermissionService, args: PermissionArguments
+) -> dict[UserRole, list[PermissionConfig]]:
     """Role-based configuration for the `/users/*` Stemmarest endpoint."""
     base_config = [
         PermissionConfig(
